refactor: move Z-Wave JS tests into dedicated submodule
Mirror the structure already used for Matter tests by moving Z-Wave JS
provider tests into tests/providers/zwave_js/. This includes the test
file, conftest fixtures, and JSON fixture files.
