+++ tests/providers/zwave_js/__init__.py
@@ -1,0 +1,1 @@
+"""Z-Wave JS provider tests."""
