+++ tests/providers/zwave_js/fixtures/controller_node_state.json
