+++ tests/providers/zwave_js/fixtures/lock_schlage_be469_state.json
