+++ tests/providers/zwave_js/fixtures/lock_schlage_be469_v2_state.json
